refactor: extracted home
